feat(pakcage.js/japafile.js/util.spec.js): set the suit test & wrote first test (#5)
Added settings to write our future tests. Implemented one simple test for camelize function just to
show how this works. We use japa because Virk says it is more suitable for testing AdonisJs. To test
the file, the naive way is to just run: 'node path/to/spec/js/file'